+++ src/routes/roomGroup/context/createGroupContext.ts
@@ -29,6 +29,7 @@ import { SpinalContext, SpinalNode, SpinalGraphService } from 'spinal-env-viewer
 import { ROOM_TYPE } from 'spinal-env-viewer-context-geographic-service/build/constants'
 import { getProfileId } from '../../../utilities/requestUtilities';
 import { ISpinalAPIMiddleware } from '../../../interfaces';
+import { awaitSync } from '../../../utilities/awaitSync';
 module.exports = function (logger, app: express.Express, spinalAPIMiddleware: ISpinalAPIMiddleware) {
   /**
  * @swagger
@@ -51,6 +52,10 @@ module.exports = function (logger, app: express.Express, spinalAPIMiddleware: IS
  *             properties:
  *                contextName:
  *                 type: string
+ *                contextColor:
+ *                 type: string
+ *                contextIcon:
+ *                 type: string
  *     responses:
  *       200:
  *         description: Create Successfully
@@ -69,19 +74,32 @@ module.exports = function (logger, app: express.Express, spinalAPIMiddleware: IS
       const graph = await spinalAPIMiddleware.getGraph();
       await SpinalGraphService.setGraph(graph);
 
+      const contextExist = await graph.getContext(req.body.contextName)
+      if (contextExist) {
+        return res.status(400).send("Context name already exists")
+      }
+  
       const context = await groupManagerService.createGroupContext(req.body.contextName, ROOM_TYPE)
+      if(req.body.contextColor){
+        context.info.add_attr({color : req.body.contextColor})
+      }
+      if(req.body.contextIcon){
+        context.info.add_attr({icon : req.body.contextIcon})
+      }
+      await awaitSync(context); // Wait for the _server_id to be assigned by hub
+      //userGraph.addContext(context);
 
-      userGraph.addContext(context);
-
-      res.status(200).json({
+      return res.status(200).json({
         name: context.getName().get(),
         staticId: context.getId().get(),
         dynamicId: context._server_id,
-        type: context.getType().get()
+        type: context.getType().get(),
+        icon: context.info.icon?.get(),
+        color: context.info.color?.get()
       });
     } catch (error) {
       if (error.code && error.message) return res.status(error.code).send(error.message);
-      res.status(500).send(error.message)
+      return res.status(400).send(error.message)
     }
   })