+++ src/routes/roomGroup/group/createGroup.ts
@@ -28,6 +28,8 @@ import groupManagerService from "spinal-env-viewer-plugin-group-manager-service"
 import { SpinalContext, SpinalNode, SpinalGraphService } from 'spinal-env-viewer-graph-service'
 import { getProfileId } from '../../../utilities/requestUtilities';
 import { ISpinalAPIMiddleware } from '../../../interfaces';
+import { awaitSync } from '../../../utilities/awaitSync';
+
 module.exports = function (logger, app: express.Express, spinalAPIMiddleware: ISpinalAPIMiddleware) {
   /**
  * @swagger
@@ -62,11 +64,12 @@ module.exports = function (logger, app: express.Express, spinalAPIMiddleware: IS
  *             type: object
  *             required:
  *               - groupName
- *               - colorName
  *             properties:
  *                groupName:
  *                 type: string
- *                colorName:
+ *                groupColor:
+ *                 type: string
+ *                groupIcon:
  *                 type: string
  *     responses:
  *       200:
@@ -86,22 +89,35 @@ module.exports = function (logger, app: express.Express, spinalAPIMiddleware: IS
       //@ts-ignore
       SpinalGraphService._addNode(category)
 
-      if (context instanceof SpinalContext && category.belongsToContext(context)) {
-        if (context.getType().get() === "geographicRoomGroupContext") {
-          groupManagerService.addGroup(context.getId().get(), category.getId().get(), req.body.groupName, req.body.colorName)
+      if(!(context instanceof SpinalContext)) return res.status(400).send("contextId does not refer to a SpinalContext");
+      if(!(category.belongsToContext(context))) return res.status(400).send("categoryId does not belong to context provided");
+      if(context.getType().get() !== 'geographicRoomGroupContext') return res.status(400).send("contextId is not a geographicRoomGroupContext");
 
-        } else {
-          res.status(400).send("node is not type of geographicRoomGroupContext ");
-        }
-      } else {
-        res.status(400).send("category not found in context");
-      }
+      const group = await groupManagerService.addGroup(
+        context.getId().get(),
+        category.getId().get(),
+        req.body.groupName,
+        req.body.groupColor,
+        req.body.groupIcon
+      )
 
+      await awaitSync(group);
+
+      return res.status(200).json({
+        name: group.getName().get(),
+        staticId: group.getId().get(),
+        dynamicId: group._server_id,
+        type: group.getType().get(),
+        icon: group.info.icon?.get(),
+        color : group.info.color?.get()
+      });
+
+      
+      
     } catch (error) {
       if (error.code && error.message) return res.status(error.code).send(error.message);
-      res.status(400).send(error.message)
+      return res.status(400).send(error.message)
     }
-    res.json();
   })
 
 }